% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{find_stab}
\alias{find_stab}
\title{The subspace stability of a column space}
\usage{
find_stab(base_lst, Z)
}
\arguments{
\item{base_lst}{a list with length 2*B, where each element gives the n by n projection matrix \eqn{P_{X_S^l}} from the l-th subsample}

\item{Z}{a matrix with linearly independent columns}
}
\value{
the subspace stability value of the column space of Z
}
\description{
The subspace stability of a column space
}