% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{FSSS}
\alias{FSSS}
\title{FSSS algorithm}
\usage{
FSSS(X, bags, alpha = 0.9, type = "greedy", verbose = 1)
}
\arguments{
\item{X}{the n by p design matrix}

\item{bags}{a list returned from l0_subsampling}

\item{alpha}{a value in (0,1), the threshold for stability with default value 0.9}

\item{type}{a character, either "greedy" (greedy algorithm) or "any-path"}

\item{verbose}{0 or 1, logical value for outputting the selection path}
}
\value{
a vector representing the selection set
}
\description{
Algorithm 1. A function that implements the FSSS algorithm
}