% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gen_block}
\alias{gen_block}
\title{Generate design matrix with complex dependency}
\usage{
gen_block(n, nb, b, eta = 0.1)
}
\arguments{
\item{n}{sample size}

\item{nb}{number of blocks}

\item{b}{number of parents in each block}

\item{eta}{standard deviation of the perturbation direction}
}
\value{
an n by nb*(b+1) matrix
}
\description{
A function that generates a design matrix with complex dependency structures.
The parent features follow N(0,1), and perturbation direction follows N(0, eta^2).
}