% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gen_clust}
\alias{gen_clust}
\title{Generate cluster design matrix}
\usage{
gen_clust(n, s, d, eta = 0.1)
}
\arguments{
\item{n}{sample size}

\item{s}{number of clusters}

\item{d}{number of features in each cluster}

\item{eta}{standard deviation of the perturbation direction}
}
\value{
an n by s*d matrix
}
\description{
A function that generates a design matrix with cluster structures.
The representative feature follows N(0,1), and perturbation direction follows
N(0, eta^2).
}