% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{l0_subsampling}
\alias{l0_subsampling}
\title{Subsampling using l0 base procedures}
\usage{
l0_subsampling(X, y, s0, num_bags = 100)
}
\arguments{
\item{X}{the n by p design matrix}

\item{y}{the length n response vector}

\item{s0}{an integer, tuning parameter for l0: the number of selected features}

\item{num_bags}{number of complementary subsample pairs B}
}
\value{
a list:
\describe{
\item{Selset}{a \eqn{2*B} by p matrix, where each row gives the estimated coefficients of each subsample}
\item{Pavg}{an n by n matrix, representing the Pavg matrix}
\item{base_lst}{a list with length \eqn{2*B}, where each element gives the n by n projection matrix \eqn{P_{X_S^l}} from the l-th subsample}
}
}
\description{
Subsampling using l0 base procedures
}