% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{search_radarchart}
\alias{search_radarchart}
\title{Radar chart of substitute (S1, S2)}
\usage{
search_radarchart(S1, S2, X, y, Selection_set, topk = 2)
}
\arguments{
\item{S1}{a vector representing set S1}

\item{S2}{a vector representing set S2}

\item{X}{the n by p design matrix}

\item{y}{the length n response vector}

\item{Selection_set}{a list where each element gives a selection set}

\item{topk}{an integer specifying the maximal number of subsets shown on the plot}
}
\description{
Both S1 and S2 need to have size at most 2.
The radar chat shows the most degenerate \code{topk} subsets of S1 and S2 respectively
}