% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{stability}
\alias{stability}
\title{The subspace stability of a selection set}
\usage{
stability(X, S, bags)
}
\arguments{
\item{X}{the n by p design matrix}

\item{S}{a vector representing selection set}

\item{bags}{a list returned from l0_subsampling}
}
\value{
the subspace stability value of S
}
\description{
Definition 2
}