% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{subs_search_k2}
\alias{subs_search_k2}
\title{Substitute searching algorithm}
\usage{
subs_search_k2(
  X,
  y,
  Selection_set,
  bags,
  alpha,
  tau0 = 0.8,
  tau1 = 0.5,
  tau2 = 0.8,
  maxreturn = 10,
  verbose = 1
)
}
\arguments{
\item{X}{the n by p design matrix}

\item{y}{the length n response vector}

\item{Selection_set}{a list where each element gives a selection set}

\item{bags}{a list returned from l0_subsampling}

\item{alpha}{a number, the threshold for stability}

\item{tau0}{a number, the threshold for substitutability (tau) metric}

\item{tau1}{a number, the threshold for feature perturbation (nabla tau) metric}

\item{tau2}{a number, threshold for degeneracy (tri tau) metric}

\item{maxreturn}{an integer, the maximum number of requested substitutes}

\item{verbose}{0 or 1, logical value for outputting the searching progress}
}
\value{
a data frame for all substitutes found
\describe{
\item{S1}{name of subsets S1}
\item{S2}{name of subsets S2}
\item{tau}{the tau metric of S1 and S2 w.r.t. the selection sets}
\item{nabla_tau}{the nabla_tau metric of S1 and S2 w.r.t. the selection sets}
\item{tri_tau}{the tri tau metric of S1 and S2 w.r.t. the selection sets}
}
}
\description{
Algorithm 2. The function only searches for subsets with size (k) at most 2.
}