% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{substab-package}
\alias{substab}
\alias{substab-package}
\title{A Package for Stable Feature Subspace Selection}
\description{
This package implements the proposals in the paper, 'Quantifying Uncertainty
and Stability Among Highly Correlated Predictors: A Subspace Perspective.' It
is focused on feature selection in the setting where features are highly
correlated.  It addresses the challenge of feature selection by shifting the
perspective from selecting certain features to selecting a certain subspace
of R^n.
}
\author{
\strong{Maintainer}: Xiaozhu Zhang \email{xzzhang@uw.edu}

Other contributors:
\itemize{
  \item Jacob Bien [contributor]
  \item Armeen Taeb [contributor]
}

}
\keyword{internal}