% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{tau}
\alias{tau}
\title{Substitutability metric of S1 and S2 w.r.t S0}
\usage{
tau(X, y, S1, S2, S0)
}
\arguments{
\item{X}{the n by p design matrix}

\item{y}{the length n response vector}

\item{S1}{a vector representing subset S1}

\item{S2}{a vector representing subset S2}

\item{S0}{a vector representing subset S0}
}
\value{
the substitutability (tau) metric
}
\description{
Definition 3
}