% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{tau_wrt_S}
\alias{tau_wrt_S}
\title{Substitutability metric of S1 and S2 w.r.t a collection of selection sets}
\usage{
tau_wrt_S(X, y, S1, S2, Selection_set)
}
\arguments{
\item{X}{the n by p design matrix}

\item{y}{the length n response vector}

\item{S1}{a vector representing subset S1}

\item{S2}{a vector representing subset S2}

\item{Selection_set}{a list where each element gives a selection set}
}
\value{
the minimum tau metric of S1 and S2 wrt to all selection sets
}
\description{
line (iii) in Algorithm 2
}