% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{tri_tau_wrt_S_radar}
\alias{tri_tau_wrt_S_radar}
\title{Degeneracy metric of S (in the substitutes (S, S_)) and its subset w.r.t a collection of selection sets}
\usage{
tri_tau_wrt_S_radar(X, y, j, S, S_, Selection_set)
}
\arguments{
\item{X}{the n by p design matrix}

\item{y}{the length n response vector}

\item{j}{a vector representing a subset of S}

\item{S}{a vector representing set S}

\item{S_}{a vector representing set S_}

\item{Selection_set}{a list where each element gives a selection set}
}
\value{
the degeneracy metric of j and S w.r.t the selection sets
}
\description{
Degeneracy metric of S (in the substitutes (S, S_)) and its subset w.r.t a collection of selection sets
}